requirements.txt
# Core ML and NLP
torch>=2.0.0
transformers>=4.30.0
numpy>=1.24.0
pandas>=2.0.0

# OCR Libraries
pytesseract>=0.3.10
paddlepaddle>=2.5.0
paddleocr>=2.7.0
Pillow>=10.0.0
opencv-python>=4.8.0

# Document Processing
PyPDF2>=3.0.0
pdf2image>=1.16.0
python-docx>=0.8.11

# NLP Libraries
spacy>=3.6.0
nltk>=3.8.0
scikit-learn>=1.3.0

# API and Web Framework
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# Utilities
python-dotenv>=1.0.0
click>=8.1.0
tqdm>=4.65.0
loguru>=0.7.0

# Configuration
pyyaml>=6.0
